Solved issue #125. Updated function  to throw errors in case of not table given.
